Extract CompactionOnEc2Resources
